Separate compatibility functions from __init__.py
